Remove chain_id from config and simplify some complexities.
