+++ ocean_lib/test/test_config.py
@@ -18,7 +18,6 @@ def test_metadataCacheUri_config_key():
     config_dict = {
         "METADATA_CACHE_URI": "http://ItWorked.com",
         "RPC_URL": GANACHE_URL,
-        "CHAIN_ID": 8996,
         "GAS_LIMIT": GAS_LIMIT_DEFAULT,
         "BLOCK_CONFIRMATIONS": 0,
         "TRANSACTION_TIMEOUT": 10 * 60,  # 10 minutes
@@ -36,7 +35,6 @@ def test_incomplete():
     `metadataCacheUri` config dict key when created via the Ocean __init__"""
     config_dict = {
         "METADATA_CACHE_URI": "http://ItWorked.com",
-        "RPC_URL": GANACHE_URL,
         "TRANSACTION_TIMEOUT": "some string",
     }
 
@@ -44,5 +42,5 @@ def test_incomplete():
         Ocean(config_dict=config_dict)
 
     exception_response = json.loads(exception_info.value.args[0])
-    assert exception_response["CHAIN_ID"] == "required"
+    assert exception_response["RPC_URL"] == "required"
     assert exception_response["TRANSACTION_TIMEOUT"] == "must be int"
